Implement Test buttons to give variables values.
The instructions weren't clear on what display should show when
entering a program. It seemed to suggest that variables are 0 until
you "run the program". I took that literally. While you are entering
a program everything is 0. At any time you can click one of the Test
buttons to assign variables values and run the program. A user
is considered to be entering a program as soon as one variable is used.